requirements.txt
Flask==2.3.2
Flask-Caching==2.0.2
requests==2.31.0
pycryptodome==3.19.0
colorama==0.4.6
protobuf==4.25.1  # <-- تم تحديث الإصدار
serverless-wsgi==3.0.0  # <-- تمت إضافته